Alert and validation, login and registration

+++ Banking-frontend/src/app/loginpage/loginpage.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { Customer } from '../app-model/customer';
 import { Login } from '../app-model/login';
 import { CustomerService } from '../customer.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-loginpage',
@@ -20,22 +21,38 @@ export class LoginpageComponent{
     this.router.navigateByUrl('');
   }
 
-  loginCheck(){
-    console.log(this.login);
-    this.service.login(this.login).subscribe(response =>{
-     // alert(JSON.stringify(response));
-      console.log(response);
-      if(response.status == true){
-        console.log(this.login.customerId);
-        let customerId = response.customerId;
-        //alert(JSON.stringify(customerId));
-        sessionStorage.setItem('customerId',String(customerId));
-        this.router.navigate(['userDashboard']);
-      }
-      else{
-        this.message = response.message;
-      }
-    })
-  }
+  loginCheck(event){
+    if(this.login.customerId == null || this.login.loginPassword == null){
+      Swal.fire({
+        title: "Error",
+        text: "Login id and password required",
+        icon: "error",
+        confirmButtonText: "Okay"
+      });
+    }
+    else{
+      console.log(this.login);
+      this.service.login(this.login).subscribe(response =>{
+      // alert(JSON.stringify(response));
+        console.log(response);
+        if(response.status == true){
+          console.log(this.login.customerId);
+          let customerId = response.customerId;
+          //alert(JSON.stringify(customerId));
+          sessionStorage.setItem('customerId',String(customerId));
+          Swal.fire({
+            title: "Successful",
+            text: "Login Successful",
+            icon: "success",
+            confirmButtonText: "Okay"
+          });
+          this.router.navigate(['userDashboard']);
+        }
+        else{
+          this.message = response.message;
+        }
+      })
+    }
+}
 
 }
